+++ Http/WsseRequestMatcher.php
@@ -2,28 +2,17 @@
 namespace Oka\ApiBundle\Http;
 
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\RequestMatcherInterface;
 
 /**
  * 
  * @author Cedrick Oka Baidai <okacedrick@gmail.com>
  * 
  */
-class WsseRequestMatcher implements RequestMatcherInterface
+class WsseRequestMatcher extends HostRequestMatcher
 {
-	/**
-	 * @var HostRequestMatcher $hostMatcher
-	 */
-	private $hostMatcher;
-	
-	public function __construct(HostRequestMatcher $hostMatcher)
-	{
-		$this->hostMatcher = $hostMatcher;
-	}
-	
 	public function matches(Request $request)
 	{
-		if (false === $this->hostMatcher->matches($request)) {
+		if (false === parent::matches($request)) {
 			return false;
 		}